+++ src/app/message_bridge/process/message_bridge_process.py
@@ -1,11 +1,11 @@
 from common.event_bus.stream_bus import StreamBus
-from common.process.bus_process import BusProcess
-from common.process.interfaces import IStreamBusProcess
+from common.process.imdg_bus_process import ImdgBusProcess
+from common.process.interfaces import IBusProcess
 from protocol.message.packet import E_PROTOCOL_MESSAGE_DIRECTION, Packet
 from protocol.protocol_meta import E_PROTOCOL_ID, ProtocolMeta
 
 
-class MessageBridgeProcess(BusProcess, IStreamBusProcess):
+class MessageBridgeProcess(ImdgBusProcess):
     def __init__(self, app_name, process_name):
         super().__init__(app_name, process_name)
         self._stream_bus: StreamBus | None = None
@@ -18,7 +18,7 @@ class MessageBridgeProcess(BusProcess, IStreamBusProcess):
         self._stream_bus.publish(message)
 
     @staticmethod
-    def playable_list_request(process: IStreamBusProcess, packet: Packet):
+    def playable_list_request(process: IBusProcess, packet: Packet):
         from process_category.enum_category import E_CATE
         from protocol.protocol_owner import ProtocolOwner
         sender = ProtocolOwner.build(E_CATE.MESSAGE_BRIDGE, E_CATE.E_MESSAGE_BRIDGE.E_COMMON.MESSAGE_BRIDGE)
